refactor: renamed state to store
