+++ tsconfig.json
@@ -27,8 +27,8 @@
             "@utils/*": ["utils/*"],
             "@test-utils": ["utils/test-utils"],
             "@styled": ["utils/styled"],
-            "@state": ["state"],
-            "@state/*": ["state/*"],
+            "@store": ["store"],
+            "@store/*": ["store/*"],
             "@routes": ["routes"],
             "@routes/*": ["routes/*"],
             "@pages": ["pages"],
